requirements.txt
flask
discord.py
requests
aiohttp
python-dotenv
